Change TSValid reporting to logging rather than stdin
This makes tsvalid more flexible for use in other toolkit.

Also, previously, we have printed a filename in the beginning of the output string, which makes no sense when the input is a stream. We only print a filename now if there is one.

The output is thoroughly tested and works as expected.